Add the pgdev.sh script
This script is the core of the pgdev tool. It handles most of the sub-
commands described in the README and some new ones:

- new
- setup
- init
- create
- start
- stop
- restart
- status
- list
- help

WIP commands:
- delete
- default
- configure
- build
- test
- psql
- logs
- conf

The 'switch' command is implemented to just raise an error message.
This command should be implemented directly in the wrapper functions for
the different shells (bash, zsh, fish, etc.), since it needs to modify
environment variables.