Add base template
This is the base template that can be used when running 'pgdev create'.
This template installs postgresql and all contrib packages from source.

A template is made up of multiple steps of .configure.sh and .build.sh
that will be run in sorted order by the pgdev script. For this reason it
is good practice to number the steps as it done here. In theory this is
not needed here as there is only 1 configure + build step, but for more
complex templates this is needed, so it is done here for consistency.

The files of a single step should all have the same base name. For
example, 01_postgres.configure.sh and 01_postgresql.build.sh won't work.

Additionally, each step can also have a .test.sh file that is used to
run tests (make installcheck, ctest, etc.) for this step. This is not
run automatically by 'pgdev create', but can be run manually using
'pgdev test <test_script>'.

For ease of maintenance, all the configure/build/test scripts are stored
in the blueprints/ folder. The template folder should only contain
relative symlinks to the blueprint files they use. This also allows the
same blueprint file to be used at different stages of different templates
as two differently named symlinks can point to the same blueprint file.

In the future, we should add a workflow action to check that all the
symlinks in the templates directory are valid and point to existing
files in the blueprints directory.